refactor(directory): add relevant component together in folder
