+++ lib/pages/home/home_view.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:one_ai/utils/components/app_bar_component.dart';
-import 'package:one_ai/utils/components/app_drawer.dart';
+import 'package:one_ai/utils/components/app_bar/app_bar_component.dart';
+import 'package:one_ai/utils/components/drawer/app_drawer.dart';
 import 'package:one_ai/utils/components/home_tile.dart';
 import 'package:one_ai/utils/components/logo_tile.dart';
-import 'package:one_ai/utils/components/prompt_input.dart';
+import 'package:one_ai/utils/components/prompt/prompt_input.dart';
 import 'package:one_ai/utils/constants/app_colors.dart';
 import 'package:one_ai/utils/constants/app_spacing.dart';
 import 'package:one_ai/utils/constants/app_text_styles.dart';
