+++ lib/utils/components/drawer/app_drawer.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:one_ai/model/drawer_action_model.dart';
-import 'package:one_ai/utils/components/drawer_action.dart';
-import 'package:one_ai/utils/components/drawer_action_tile.dart';
-import 'package:one_ai/utils/components/drawer_section_tile.dart';
+import 'package:one_ai/utils/components/drawer/drawer_action_tile.dart';
+import 'package:one_ai/utils/components/drawer/drawer_section_tile.dart';
 import 'package:one_ai/utils/components/logo_tile.dart';
 import 'package:one_ai/utils/constants/app_colors.dart';
+import 'package:one_ai/utils/constants/app_radius.dart';
 import 'package:one_ai/utils/constants/app_spacing.dart';
 import 'package:one_ai/utils/constants/app_text_styles.dart';
 
@@ -46,7 +46,16 @@ class AppDrawer extends StatelessWidget {
                       fontWeight: FontWeight.w400,
                     ),
                   ),
-                  DrawerAction(
+                  Container(
+                    padding: EdgeInsets.symmetric(horizontal: 16, vertical: 10),
+                    decoration: BoxDecoration(
+                      gradient: LinearGradient(
+                        colors: [Color(0xFFE0E7FF), Color(0xFFF3E8FF)],
+                        begin: Alignment.topLeft,
+                        end: Alignment.bottomRight,
+                      ),
+                      borderRadius: AppRadius.radiusCirular,
+                    ),
                     child: Row(
                       children: [
                         Icon(
@@ -212,5 +221,3 @@ class AppDrawer extends StatelessWidget {
     );
   }
 }
-
-
